regiohelden/requirements.txt
Django
psycopg2
django-allauth
django-localflavor
djangorestframework